Fixed Typos in classConstructors.md
Fixed an issue in the "Constructor Parameters" and the "Class Destructors" code examples where the parameter names to the constructor were swapped. I've left the input to the constructor the same, but I've swapped the order in the print statement so that it gives the same output. Also fixed a typo where the expected output of the print statement included commas, providing multiple parameters to `print()` only separates them a space, not a comma and a space.